add modules task implementation
